splitting redis-mock and redis-client into separate files and extracting the index.js
