+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.55.1",
   "description": "Redis client mock object for unit testing",
   "author": "Kristian Faeldt <kristian.faeldt@gmail.com>",
-  "main": "./lib/redis-mock",
+  "main": "./lib/index",
   "license": "MIT",
   "repository": {
     "type": "git",
